Initial Commit
This is the first working version of the code.  it will transform a
flat csv file with with a unique identifier for each row, and a
reference to each parent relationship represented in another row, where
the parent relationship to each child is represented by a parent id
reference for each record in every row where the record has a parent
relationship.